+++ src/gltf/sqlite.ts
@@ -1,15 +1,17 @@
 import * as sqlite3 from 'sqlite3';
 import * as gltf from './schema';
 import { isUndefined } from 'util';
+import { PropDbReader } from '../common/propdb-reader';
 
 /**
  * Serializes glTF manifest into sqlite database on disk.
  * @async
  * @param {gltf.Gltf} gltf Original manifest.
  * @param {string} sqlitePath Path to local file where the sqlite database should be stored.
+ * @param {PropDbReader} [pdb] Optional property reader for properties to be merged into the sqlite manifest.
  * @returns {Promise<void>} Promise that resolves after the database is created.
  */
-export async function serialize(gltf: gltf.GlTf, sqlitePath: string): Promise<void> {
+export async function serialize(gltf: gltf.GlTf, sqlitePath: string, pdb?: PropDbReader): Promise<void> {
     const db = await openDatabase(sqlitePath);
     await serializeNodes(db, gltf);
     await serializeMeshes(db, gltf);
@@ -19,16 +21,25 @@ export async function serialize(gltf: gltf.GlTf, sqlitePath: string): Promise<vo
     await serializeMaterials(db, gltf);
     await serializeTextures(db, gltf);
     await serializeImages(db, gltf);
+    if (pdb) {
+        await serializeProperties(db, gltf, pdb);
+    }
     await closeDatabase(db);
 }
 
 /**
  * Deserializes glTF manifest from sqlite database on disk.
  * @param {string} sqlitePath Path to local file the sqlite database should be read from.
- * @param {string} [filter] Optional body for a SQL WHERE-clause to be used for filtering dbids
- * in the deserialized glTF, for example, 'dbid > 2000 AND dbid < 4000'.
+ * @param {string | number[]} [filter] Optional filter of object IDs to include in the output gltf.
+ * Can be either a sqlite query returning a list of dbIDs, or an array of dbIDs.
+ * @example
+ * const gltf = await deserialize('./data/model/manifest.sqlite', 'SELECT dbid FROM nodes WHERE dbid >= 1000 AND dbid <= 2000');
+ * @example
+ * const gltf = await deserialize('./data/model/manifest.sqlite', 'SELECT dbid FROM properties WHERE name = "Material" AND value LIKE "%Concrete%"');
+ * @example
+ * const gltf = await deserialize('./data/model/manifest.sqlite', [100, 101, 102, 103]);
  */
-export async function deserialize(sqlitePath: string, filter?: string): Promise<gltf.GlTf> {
+export async function deserialize(sqlitePath: string, filter?: string | number[]): Promise<gltf.GlTf> {
     let gltf: gltf.GlTf = {
         asset: {
             version: '2.0',
@@ -49,7 +60,7 @@ export async function deserialize(sqlitePath: string, filter?: string): Promise<
         scene: 0
     };
     const db = await openDatabase(sqlitePath);
-    const { meshSet } = await deserializeNodes(db, gltf, isUndefined(filter) ? '' : 'WHERE ' + filter);
+    const { meshSet } = await deserializeNodes(db, gltf, filter);
     const { materialSet, accessorSet} = await deserializeMeshes(db, gltf, meshSet);
     const { bufferViewSet } = await deserializeAccessors(db, gltf, accessorSet);
     const { bufferSet } = await deserializeBufferViews(db, gltf, bufferViewSet);
@@ -84,6 +95,30 @@ function closeDatabase(db: sqlite3.Database): Promise<void> {
 function run(db: sqlite3.Database, sql: string): Promise<void> {
     return new Promise(function (resolve, reject) {
         db.run(sql, function (err: Error) {
+            if (err) {
+                reject(err);
+            } else {
+                resolve();
+            }
+        });
+    });
+}
+
+async function serializeProperties(db: sqlite3.Database, gltf: gltf.GlTf, pdb: PropDbReader): Promise<void> {
+    // Collect list of unique dbids already in the database
+    const dbids = await readDbIds(db);
+
+    // Add properties for each dbid
+    await run(db, 'CREATE TABLE properties (dbid INTEGER, name TEXT, value TEXT)');
+    let stmt = db.prepare('INSERT INTO properties VALUES (?, ?, ?)');
+    for (const dbid of dbids) {
+        const props = pdb.getProperties(dbid);
+        for (const key of Object.keys(props)) {
+            stmt.run(dbid, key, props[key]);
+        }
+    }
+    return new Promise(function (resolve, reject) {
+        stmt.finalize(function (err) {
             if (err) {
                 reject(err);
             } else {
@@ -283,14 +318,14 @@ async function serializeNodes(db: sqlite3.Database, gltf: gltf.GlTf): Promise<vo
     }
 }
 
-function deserializeNodes(db: sqlite3.Database, gltf: gltf.GlTf, filter: string): Promise<{ count: number; meshSet: Set<number> }> {
+function deserializeNodes(db: sqlite3.Database, gltf: gltf.GlTf, filter?: string | number[]): Promise<{ count: number; meshSet: Set<number> }> {
     const meshSet = new Set<number>(); // Set of all mesh IDs to be returned by this function's promise
     const nodes = gltf.nodes as gltf.Node[];
     const scene = (gltf.scenes as gltf.Scene[])[0];
-    const query = `
+    let query = `
         SELECT id, dbid, mesh_id, matrix_json AS mtx, translation_x AS tx, translation_y AS ty, translation_z AS tz, scale_x AS sx, scale_y AS sy, scale_z AS sz, rotation_x AS rx, rotation_y AS ry, rotation_z AS rz, rotation_w AS rw
         FROM nodes
-        ${filter}
+        ${filter ? `WHERE dbid IN (${typeof filter === 'string' ? filter : filter.join(',')})` : ''}
     `;
     const onRow = (err: Error, row: any) => {
         if (err) {
@@ -678,4 +713,20 @@ function rebuildIndices(gltf: gltf.GlTf) {
     for (const image of images) {
         delete image.__id;
     }
+}
+
+function readDbIds(db: sqlite3.Database): Promise<number[]> {
+    let query = `
+        SELECT DISTINCT dbid
+        FROM nodes
+    `;
+    return new Promise(function (resolve, reject) {
+        db.all(query, function onComplete(err: Error, rows: any[]) {
+            if (err) {
+                reject(err);
+            } else {
+                resolve(rows.map(row => row.dbid));
+            }
+        });
+    });
 }